Import live.2006.05.11.tar.gz
- Cleaned up some code in "BasicTaskScheduler" that was allegedly causing
  problems on a Zaurus.
- Changed "TaskScheduler::rescheduleDelayedTask()" so that its "microseconds"
  parameter is a "int64_t" (to match "scheduleDelayedTask").
  (Thanks to David Bertrand for suggesting this.)
- Updated "ServerMediaSession:::generateSDPDescription()" to change
  "a=rtcp: unicast-reflection" to "a=rtcp-unicast: reflection", to
  reflect the latest version of the IETF's "draft-ietf-avt-rtcpssm"
  Internet-Draft.  (This is for Source-Specific Multicast sessions only.)
- Cleaned up "H263plusVideoRTPSink" slightly, to remove an unnecessary
  member variable.
- Removed support for the (experimental and non-standard) "X-MCT-TEXT"
  payload type.

+++ UsageEnvironment/UsageEnvironment.cpp
@@ -38,7 +38,7 @@ TaskScheduler::~TaskScheduler() {
 }
  
 void TaskScheduler::rescheduleDelayedTask(TaskToken& task,
-					  int microseconds, TaskFunc* proc,
+					  int64_t microseconds, TaskFunc* proc,
 					  void* clientData) {
   unscheduleDelayedTask(task);
   task = scheduleDelayedTask(microseconds, proc, clientData);
